turn streaming/processing components into module
