+++ packages/grz-common/src/grz_common/workers/submission.py
@@ -32,7 +32,8 @@ from tqdm.auto import tqdm
 
 from ..constants import TQDM_DEFAULTS
 from ..models.identifiers import IdentifiersModel
-from ..pipeline.components import Crypt4GHDecryptor, Crypt4GHEncryptor, TqdmObserver
+from ..pipeline.components import TqdmObserver
+from ..pipeline.components.crypt4gh import Crypt4GHDecryptor, Crypt4GHEncryptor
 from ..progress import DecryptionState, EncryptionState, FileProgressLogger, ValidationState
 from ..utils.crypt import Crypt4GH
 from ..validation import UserInterruptException, run_grz_check
